+++ doc/changes.md
@@ -310,6 +310,9 @@ as part of K95 at this time, the default terminal remains VT220 for now.
    - Text-mode popups now work when the cursor is in the status line
    - The cursor no longer vanishes while the host is in the status line in K95G
    - Fixed mark mode when the cursor is in the status line
+   - Fixed setting and going to bookmarks while the cursor is in the status line
+   - Fixed viewing the scrollback while the cursor is in the status line
+
 
 ## Kermit 95 v3.0 beta 7 - 27 January 2025